+++ index.html
@@ -33,6 +33,8 @@
   </style>
 </body>
 
-<script src="index.js"></script>
+<script type="module" src="index.js"></script>
+<script type="module" src="API.js"></script>
+<script type="module" src="rend.js"></script>
 
 </html>
