+++ styles.css
@@ -2,6 +2,10 @@ body {
   margin: 0;
 }
 
+.-display-none {
+  display: none;
+}
+
 .container {
   font-family: Helvetica;
   color: #ffffff;
